rgw: get frontend ssl cert and key from mon store
If ssl_certificate or ssl_key is missing from the rgw_frontends then RGW
will check the mon config-key store for them.

The keys are:
  - "rgw/frontend/cert": for the cert.
  - "rgw/frontend/key": for the key.

I'm not sure if the config-key's can be scoped like the config store. If
they can't, then there can only be 1 cert for a cluster. If more
grainularity is required. We could change these keys to something like
'rgw_frontend_{key,cert}' and then we can scope them to the client, ie
general client, realm or zone.

Signed-off-by: Matthew Oliver <moliver@suse.com>